Simple PHP/MySql application about Bookmarks
It was made for a School cource about PHP/MySQL development for Laurea School of Applied Sciences.
It was based on a tutorial from a book called PHP and MySQL web development by Luke Welling and Laura Thomson. I made my own fixes to it and structured it bit diffrently. But the concept is quite the same.